Reusable github action (first test)
