+++ modules-magic-drc/action.yml
@@ -1,0 +1,46 @@
+# Copyright 2021 Efabless Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+name: 'Skywater PDK Libraries - Run Magic DRC on All Cells'
+description: |
+  This runs DRC using Magic on all SCL GDS files.
+
+  This also ensures that GDS files can be read by Magic.
+
+  Requires having checked out the repository.
+inputs:
+  match_directories:
+    description: |
+      Match directories under <repository_root>/cells/ with this regular expression.
+      Default: . (matches all)
+    required: false
+    default: '.'
+  known_bad:
+    description: |
+      List of cell names that are known bad. known_bad cells do not contribute
+      to the failure of this task. Comma,delimited,list
+      Default: Empty string (ignores nothing).
+    required: false
+    default: ''
+runs:
+  using: docker
+  image: 'Dockerfile'
+  args:
+    - '-m'
+    - ${{ inputs.match_directories }}
+    - '-b'
+    - ${{ inputs.known_bad }}
+
