+++ modules-magic-drc/pdk_subset/libs.tech/magic/sky130A.tcl
