+++ modules-magic-drc/test_action_offline.sh
@@ -1,0 +1,26 @@
+#!/bin/sh
+# Copyright 2020 Efabless Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+# --
+# This file tests modules-magic-drc offline.
+#
+if [ $PWD != $(realpath $(dirname $(dirname $0))) ]; then
+    echo "Must be run from the root of the repository."
+    exit 64
+fi
+
+docker build -t modules-magic-drc -f modules-magic-drc/Dockerfile modules-magic-drc
+docker run --rm -tiv $PWD:/github_workspace -w /github_workspace modules-magic-drc -m ${1-.} -b ${2-none}
